// WebApp.Client/src/generated/dtos.ts
export class ProblemDetails implements IProblemDetails {
    type?: string | undefined;
    title?: string | undefined;
    status?: number | undefined;
    detail?: string | undefined;
    instance?: string | undefined;

    [key: string]: any;

    constructor(data?: IProblemDetails) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            for (var property in _data) {
                if (_data.hasOwnProperty(property))
                    this[property] = _data[property];
            }
            this.type = _data["type"];
            this.title = _data["title"];
            this.status = _data["status"];
            this.detail = _data["detail"];
            this.instance = _data["instance"];
        }
    }

    static fromJS(data: any): ProblemDetails {
        data = typeof data === 'object' ? data : {};
        let result = new ProblemDetails();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        for (var property in this) {
            if (this.hasOwnProperty(property))
                data[property] = this[property];
        }
        data["type"] = this.type;
        data["title"] = this.title;
        data["status"] = this.status;
        data["detail"] = this.detail;
        data["instance"] = this.instance;
        return data;
    }
}

export interface IProblemDetails {
    type?: string | undefined;
    title?: string | undefined;
    status?: number | undefined;
    detail?: string | undefined;
    instance?: string | undefined;

    [key: string]: any;
}
// ...
export class FamilyUserDto implements IFamilyUserDto {
    id!: number;
    user!: UserMinimalDto;
    family!: FamilyHeaderDto;
    isOwner!: boolean;
    joinedAt!: Date;
    updatedByUserId!: number;

    constructor(data?: IFamilyUserDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
        if (!data) {
            this.user = new UserMinimalDto();
            this.family = new FamilyHeaderDto();
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.user = _data["user"] ? UserMinimalDto.fromJS(_data["user"]) : new UserMinimalDto();
            this.family = _data["family"] ? FamilyHeaderDto.fromJS(_data["family"]) : new FamilyHeaderDto();
            this.isOwner = _data["isOwner"];
            this.joinedAt = _data["joinedAt"] ? new Date(_data["joinedAt"].toString()) : undefined as any;
            this.updatedByUserId = _data["updatedByUserId"];
        }
    }

    static fromJS(data: any): FamilyUserDto {
        data = typeof data === 'object' ? data : {};
        let result = new FamilyUserDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["user"] = this.user ? this.user.toJSON() : undefined as any;
        data["family"] = this.family ? this.family.toJSON() : undefined as any;
        data["isOwner"] = this.isOwner;
        data["joinedAt"] = this.joinedAt ? this.joinedAt.toISOString() : undefined as any;
        data["updatedByUserId"] = this.updatedByUserId;
        return data;
    }
}

export interface IFamilyUserDto {
    id: number;
    user: UserMinimalDto;
    family: FamilyHeaderDto;
    isOwner: boolean;
    joinedAt: Date;
    updatedByUserId: number;
}

export class UserMinimalDto implements IUserMinimalDto {
    id?: number;
    userName?: string;

    constructor(data?: IUserMinimalDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.userName = _data["userName"];
        }
    }

    static fromJS(data: any): UserMinimalDto {
        data = typeof data === 'object' ? data : {};
        let result = new UserMinimalDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["userName"] = this.userName;
        return data;
    }
}

export interface IUserMinimalDto {
    id?: number;
    userName?: string;
}

export class FamilyHeaderDto implements IFamilyHeaderDto {
    id?: number;
    createdAt?: Date;
    isShared?: boolean;
    name?: string;

    constructor(data?: IFamilyHeaderDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.createdAt = _data["createdAt"] ? new Date(_data["createdAt"].toString()) : undefined as any;
            this.isShared = _data["isShared"];
            this.name = _data["name"];
        }
    }

    static fromJS(data: any): FamilyHeaderDto {
        data = typeof data === 'object' ? data : {};
        let result = new FamilyHeaderDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["createdAt"] = this.createdAt ? this.createdAt.toISOString() : undefined as any;
        data["isShared"] = this.isShared;
        data["name"] = this.name;
        return data;
    }
}

export interface IFamilyHeaderDto {
    id?: number;
    createdAt?: Date;
    isShared?: boolean;
    name?: string;
}
// ...
export class UserHeaderDto extends UserMinimalDto implements IUserHeaderDto {
    email?: string;
    dateOfBirth?: Date;
    totalBalance?: number;

    constructor(data?: IUserHeaderDto) {
        super(data);
    }

    override init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.email = _data["email"];
            this.dateOfBirth = _data["dateOfBirth"] ? new Date(_data["dateOfBirth"].toString()) : undefined as any;
            this.totalBalance = _data["totalBalance"];
        }
    }

    static override fromJS(data: any): UserHeaderDto {
        data = typeof data === 'object' ? data : {};
        let result = new UserHeaderDto();
        result.init(data);
        return result;
    }

    override toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["email"] = this.email;
        data["dateOfBirth"] = this.dateOfBirth ? this.dateOfBirth.toISOString() : undefined as any;
        data["totalBalance"] = this.totalBalance;
        super.toJSON(data);
        return data;
    }
}

export interface IUserHeaderDto extends IUserMinimalDto {
    email?: string;
    dateOfBirth?: Date;
    totalBalance?: number;
}
// ...
export class LimitDto implements ILimitDto {
    id?: number;
    user?: UserHeaderDto;
    description?: string | undefined;
    maxAmount?: number;
    currentAmount?: number;
    startDate?: Date;
    endDate?: Date;
    createdAt?: Date;
    createdByUserId?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    includeIncome?: boolean;

    constructor(data?: ILimitDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.user = _data["user"] ? UserHeaderDto.fromJS(_data["user"]) : undefined as any;
            this.description = _data["description"];
            this.maxAmount = _data["maxAmount"];
            this.currentAmount = _data["currentAmount"];
            this.startDate = _data["startDate"] ? new Date(_data["startDate"].toString()) : undefined as any;
            this.endDate = _data["endDate"] ? new Date(_data["endDate"].toString()) : undefined as any;
            this.createdAt = _data["createdAt"] ? new Date(_data["createdAt"].toString()) : undefined as any;
            this.createdByUserId = _data["createdByUserId"];
            this.isActive = _data["isActive"];
            this.isDeleted = _data["isDeleted"];
            this.includeIncome = _data["includeIncome"];
        }
    }

    static fromJS(data: any): LimitDto {
        data = typeof data === 'object' ? data : {};
        let result = new LimitDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["user"] = this.user ? this.user.toJSON() : undefined as any;
        data["description"] = this.description;
        data["maxAmount"] = this.maxAmount;
        data["currentAmount"] = this.currentAmount;
        data["startDate"] = this.startDate ? formatDate(this.startDate) : undefined as any;
        data["endDate"] = this.endDate ? formatDate(this.endDate) : undefined as any;
        data["createdAt"] = this.createdAt ? this.createdAt.toISOString() : undefined as any;
        data["createdByUserId"] = this.createdByUserId;
        data["isActive"] = this.isActive;
        data["isDeleted"] = this.isDeleted;
        data["includeIncome"] = this.includeIncome;
        return data;
    }
}

export interface ILimitDto {
    id?: number;
    user?: UserHeaderDto;
    description?: string | undefined;
    maxAmount?: number;
    currentAmount?: number;
    startDate?: Date;
    endDate?: Date;
    createdAt?: Date;
    createdByUserId?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    includeIncome?: boolean;
}

export class CreateLimitDto implements ICreateLimitDto {
    userId?: number;
    maxAmount?: number;
    startDate?: Date;
    endDate?: Date;
    includeIncome?: boolean;
    isActive?: boolean;
    description?: string | undefined;

    constructor(data?: ICreateLimitDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.userId = _data["userId"];
            this.maxAmount = _data["maxAmount"];
            this.startDate = _data["startDate"] ? new Date(_data["startDate"].toString()) : undefined as any;
            this.endDate = _data["endDate"] ? new Date(_data["endDate"].toString()) : undefined as any;
            this.includeIncome = _data["includeIncome"];
            this.isActive = _data["isActive"];
            this.description = _data["description"];
        }
    }

    static fromJS(data: any): CreateLimitDto {
        data = typeof data === 'object' ? data : {};
        let result = new CreateLimitDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["userId"] = this.userId;
        data["maxAmount"] = this.maxAmount;
        data["startDate"] = this.startDate ? formatDate(this.startDate) : undefined as any;
        data["endDate"] = this.endDate ? formatDate(this.endDate) : undefined as any;
        data["includeIncome"] = this.includeIncome;
        data["isActive"] = this.isActive;
        data["description"] = this.description;
        return data;
    }
}

export interface ICreateLimitDto {
    userId?: number;
    maxAmount?: number;
    startDate?: Date;
    endDate?: Date;
    includeIncome?: boolean;
    isActive?: boolean;
    description?: string | undefined;
}